# postman/specs/api-contract/api-contract.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "api-contract",
    "version": "1.0.0",
    "description": "Colección de requests para testing manual de la API de QuizForge"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/preguntas": {
      "post": {
        "summary": "Crear Pregunta Multiple Choice",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "Bearer {{jwt_token}}"
          }
        ],
        "responses": {},
        "tags": [
          "Preguntas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipo": {
                    "type": "string"
                  },
                  "texto": {
                    "type": "string"
                  },
                  "opciones": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "utc-millisec"
                    }
                  },
                  "respuestaCorrecta": {
                    "type": "string",
                    "format": "utc-millisec"
                  },
                  "examenId": {
                    "type": "integer"
                  }
                }
              },
              "example": {
                "tipo": "MULTIPLE_CHOICE",
                "texto": "¿En qué año cayó el muro de Berlín?",
                "opciones": [
                  "1987",
                  "1989",
                  "1991",
                  "1993"
                ],
                "respuestaCorrecta": "1989",
                "examenId": 2
              }
            }
          }
        }
      }
    },
    "/api/preguntas/masa": {
      "post": {
        "summary": "Crear Preguntas en Masa",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "Bearer {{jwt_token}}"
          }
        ],
        "responses": {},
        "tags": [
          "Preguntas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "tipo": {
                      "type": "string"
                    },
                    "texto": {
                      "type": "string"
                    },
                    "opciones": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "respuestaCorrecta": {
                      "type": "string"
                    },
                    "examenId": {
                      "type": "integer"
                    }
                  }
                }
              },
              "example": [
                {
                  "tipo": "MULTIPLE_CHOICE",
                  "texto": "¿Cuál es el planeta más grande del sistema solar?",
                  "opciones": [
                    "Tierra",
                    "Marte",
                    "Júpiter",
                    "Saturno"
                  ],
                  "respuestaCorrecta": "Júpiter",
                  "examenId": 1
                },
                {
                  "tipo": "VERDADERO_FALSO",
                  "texto": "El agua hierve a 100°C a nivel del mar",
                  "respuestaCorrecta": true,
                  "examenId": 1
                },
                {
                  "tipo": "DESARROLLO_DETERMINISTICO",
                  "texto": "¿Cuánto es 2 + 2?",
                  "respuestaEsperadaExacta": "4",
                  "examenId": 1
                }
              ]
            }
          }
        }
      }
    },
...
  }
}